New memory handling via range-checking/propagation functions
Replace direct access to large memory buffers for CPU and PPU memory with read/write functions that check address ranges and propagate to smaller buffers that represent the actual hardware. This simplifies the implementation of memory access side-effects (register shifts, etc.), and will make it much easier to implement mappers.

+++ src/Base.h
@@ -41,7 +41,7 @@ static_assert(sizeof(int32)==4, "Invalid type size");
 #define TO16(v8) ((uint16)(v8))
 #define TO8(v16) ((uint8)(v16 & 0x00FF))
 
-#define BIT(n) ((uint8)(1<<n))
+#define BIT(n) (1<<n)
 
 #define ARRAYSIZE(arr) (sizeof(arr)/sizeof(arr[0]))
